Update non-behavior client tests for event log format.
All pass.

Decided for now not to verify watchpoint hits - is hassle and not
generally useful (see comment in checkEvents).

Includes some fixes for go 1.20 (unnamed recvrs get compiled out),
delve server (WatchType limited max watch size to 4 bits, pass move
correctly on recursive CreateWatchpoint calls in Debugger),
delve client (pass correct watchexpr to event, fix thread id output).

None move wps except allocator_http.
Also added env var option to keep test csvs.

+++ pkg/proc/native/hwbreak_amd64.go
@@ -7,7 +7,7 @@ import (
 
 func (t *nativeThread) writeHardwareBreakpoint(addr uint64, wtype proc.WatchType, idx uint8) error {
 	return t.withDebugRegisters(func(drs *amd64util.DebugRegisters) error {
-		return drs.SetBreakpoint(idx, addr, wtype.Read(), wtype.Write(), wtype.Size())
+		return drs.SetBreakpoint(idx, addr, wtype.Read(), wtype.Write(), (int)(wtype.Size()))
 	})
 }
